% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/a.R, R/b.R, R/base.R
\name{a}
\alias{a}
\alias{b}
\alias{base}
\title{Base}
\usage{
a(x)

b(x)

base(x)
}
\description{
Desc-A

Desc-B

Description
}
\details{
Detail-A

Detail-B

Detail
}
\section{OBTW}{


TL;DR
}

\section{OBTW}{


ORLY?
}

